fix(v2): match reserve RPC percentage types

+++ energiazen-mini/supabase/migrations/20260916114500_add_v2_energy_reserve_home_view.sql
@@ -6,8 +6,8 @@ returns table (
   available boolean,
   conservative_energy_kwh double precision,
   energy_capacity_kwh double precision,
-  safety_reserve_percent integer,
-  target_reserve_percent integer
+  safety_reserve_percent double precision,
+  target_reserve_percent double precision
 )
 language sql
 security definer
